test(tui): settle strict-tilde frames on the post-highlight state; share the frame helper (#305 review)
Track 1 CR + Bot Review (both approve) findings:

- The prose component test polled on a needle already present in the
  pre-highlight frame, so a slow worker could return early with raw
  ~~删除~~ still visible. The shared helper is now predicate-based
  (captureFrameUntil) and the test settles on the post-highlight state
  (literal single tildes present, doubled tildes concealed).
- captureFrameWhen was a verbatim duplicate across the strict-tilde and
  escape suites; both now import from tests/support/markdown-frame.ts.
- TUI.md/CHANGELOG scope the plain-renderer claim to lone single-tilde
  spans (doubled-tilde cleanup and subscript rewrite still transform), and
  the subscript pre-transform note now states its actual shape (eligible
  ASCII spans: digit-bearing or single-character).
