+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - **New sessions now explicitly create the model-writable project roots at birth (issue #291, from #268 item 1).** `source_materials/`, `workspace/`, `novels/`, `reports/`, `test_runs/`, and `tmp/` are created up front instead of appearing lazily on first write, so a fresh project's first `<project_state>` reads `empty` rather than a wall of `absent`, and Workspace shows the standard folders once a session exists. Creation is best effort: a squatted or unreadable path (for example a regular file named `workspace`) only produces a transcript warning, and project state keeps reporting that root honestly as `inaccessible`. The read-only `assets` root is excluded; resumed sessions do not recreate roots you deleted, and the write tools' lazy-mkdir backstop is unchanged.
 - **Terminal-title idle and input-required markers now share the working pulse's square family (issue #290, from #268 item 7).** The static state markers move from `·` (U+00B7) and `!` (U+0021) — too light next to the working frames at tab-strip size — to `■` (U+25A0) for idle and `▣` (U+25A3) for input-required, the same square silhouette as the `◰`/`◳`/`◲`/`◱` pulse at the same single-column advance, so the three states read as one family. The working frame set, 800ms cadence, `VESICLE_REDUCED_MOTION=1` still frame, and no-write overrides are unchanged. The scheme was selected by human-observation probe (`scripts/probe/terminal-title-marker-probe.ts`).
 - **OpenTUI fork runtime advances to `0.5.10-zv5`** (from `0.5.3-zv6`; fork baseline rejoined upstream at v0.5.9 and now carries upstream v0.5.10; bare `#N` below stays host issues, upstream changes carry the `upstream #N` qualifier). From the v0.5.9 baseline: the byte-accurate layout rewrite (upstream #1428) and drag-selection inclusive-end semantics (upstream #1393) — dragging across cells includes the end cell, so the theme-text selection expectations carry a trailing space. From v0.5.10: compatible top-level markdown renderables stay mounted at stream completion instead of being destroyed and recreated while completion UI renders (upstream #1469), the final automatic frame is retried when output rejects it under backpressure instead of leaving the previous frame displayed (upstream #1457), image streaming overhead is reduced and native scrollback images are preserved above the pinned footer (upstream #1467, #1470), and the fork-built linux/win32 platform packages ship stripped native libraries with detached release symbols (upstream #1437). Validation on the bumped runtime: full local suite 2429 tests (2419 pass, 10 platform skips, 0 fail) with no expectation changes, plus the npm package consumer smoke (global/local install, audit, PTY TUI startup; following #273, issue #294).
-- **Markdown strikethrough now requires doubled tildes — lone `~text~` renders as literal text (issue #294, from #268 item 5).** Group testing flagged that prose like `~大概这样~` silently dimmed with its tildes hidden: the stock markdown parser treats any single-tilde span as strikethrough, following GFM-as-implemented (the spec's Example 491 and GitHub accept one or two tildes). Vesicle now adopts the fork's strict double-tilde flavor, matching the pandoc/remark-gfm reading and peer-agent precedent (claude-code#19251, copilot-cli#1936): `~~text~~` strikes and conceals its delimiters exactly as before, while `~text~` keeps every character visible as plain text — in prose, table cells, and while streaming. The `H~2~O` → `H₂O` subscript display transform still runs before the parser, so subscripts are unaffected, and the plain-text renderer always printed `~` literally and is unchanged. Known residual: a lone single-tilde span wrapping a doubled one (`~a ~~b~~ c~`) still dims as before — tightening that needs fork-side source filtering and stays a documented limitation. The fork's stock query stays byte-identical; the strict variant ships as a separate query asset, and a new contract test pins the capture matrix against a silent fork regression.
+- **Markdown strikethrough now requires doubled tildes — lone `~text~` renders as literal text (issue #294, from #268 item 5).** Group testing flagged that prose like `~大概这样~` silently dimmed with its tildes hidden: the stock markdown parser treats any single-tilde span as strikethrough, following GFM-as-implemented (the spec's Example 491 and GitHub accept one or two tildes). Vesicle now adopts the fork's strict double-tilde flavor, matching the pandoc/remark-gfm reading and peer-agent precedent (claude-code#19251, copilot-cli#1936): `~~text~~` strikes and conceals its delimiters exactly as before, while `~text~` keeps every character visible as plain text — in prose, table cells, and while streaming. The `H~2~O` → `H₂O` subscript display transform still runs before the parser, so subscripts are unaffected, and the plain-text renderer keeps lone single-tilde spans literal and is unchanged. Known residual: a lone single-tilde span wrapping a doubled one (`~a ~~b~~ c~`) still dims as before — tightening that needs fork-side source filtering and stays a documented limitation. The fork's stock query stays byte-identical; the strict variant ships as a separate query asset, and a new contract test pins the capture matrix against a silent fork regression.
 
 ### Fixed
