+++ host-assets/skills/vesicle-docs/references/dev-tui.md
@@ -28,7 +28,7 @@ The brand aesthetic, palette of record, motion grammar, and anti-patterns are ow
 ## Rendering
 
 - Markdown extension and LaTeX cleanup are display-only transformations. They must not mutate session records, provider messages, or artifact files.
-- Markdown strikethrough requires doubled tildes: lone `~text~` renders as literal text, intentionally stricter than GFM-as-implemented (the spec's Example 491 and GitHub both accept single tildes; pandoc and remark-gfm do not). The flavor is the fork's `highlights.strict.scm` markdown_inline parser variant registered as a default parser at TUI bootstrap — before the first tree-sitter client initializes — plus the `strikethrough="double-tilde"` option on the chat/preview markdown renderables, which also guards the marked inline path (table cells, streaming placeholders). The plain-text renderer (`VESICLE_MARKDOWN_RENDERER=plain`) always outputs `~` literally, and the `H~2~O` subscript pre-transform consumes ASCII single-tilde spans before the parser, so subscript usage never regresses. Known residual: `~a ~~b~~ c~` (a lone single-tilde span wrapping a doubled one) still dims with its outer delimiters concealed, matching the pre-strict behavior; fixing it needs worker-side source filtering in the fork. `tests/contract/tui/markdown-strict-tilde.test.ts` pins the query asset against the product oracle and `tests/component/tui/markdown-strict-tilde.test.tsx` pins both rendering paths (issue #294, from #268 item 5).
+- Markdown strikethrough requires doubled tildes: lone `~text~` renders as literal text, intentionally stricter than GFM-as-implemented (the spec's Example 491 and GitHub both accept single tildes; pandoc and remark-gfm do not). The flavor is the fork's `highlights.strict.scm` markdown_inline parser variant registered as a default parser at TUI bootstrap — before the first tree-sitter client initializes — plus the `strikethrough="double-tilde"` option on the chat/preview markdown renderables, which also guards the marked inline path (table cells, streaming placeholders). The plain-text renderer (`VESICLE_MARKDOWN_RENDERER=plain`) keeps lone single-tilde spans literal — its doubled-tilde delimiter cleanup and subscript rewrite are unchanged — and the `H~2~O` subscript pre-transform consumes eligible ASCII single-tilde spans (digit-bearing or single-character) before the parser, so subscript usage never regresses. Known residual: `~a ~~b~~ c~` (a lone single-tilde span wrapping a doubled one) still dims with its outer delimiters concealed, matching the pre-strict behavior; fixing it needs worker-side source filtering in the fork. `tests/contract/tui/markdown-strict-tilde.test.ts` pins the query asset against the product oracle and `tests/component/tui/markdown-strict-tilde.test.tsx` pins both rendering paths (issue #294, from #268 item 5).
 - Rendering cleanup must stay outside fenced code blocks and use readable static fallbacks for terminal-hostile constructs.
 - Thinking content renders separately from assistant prose, before the assistant body, with bounded or collapsible presentation.
 - Tool, Agent, validation, quality, and host-action records render from structured state rather than parsing natural-language result text.
